Add two tasks of daily coding

+++ 7.Second-Year/6.Sixth/src/HW_6_1/FirstTask.java
@@ -58,7 +58,7 @@ public class FirstTask {
         return Math.sqrt(pow(x) + pow(y));
     }
 
-    private static int pow(int n) {
+    private static long pow(int n) {
         return n * n;
     }
 }
